requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
djangorestframework==3.14.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
peewee==3.16.0
pycodestyle==2.10.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
urllib3==1.26.14
yarl==1.8.2
